pyproject.toml
[project]
name = "code8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-search-results>=2.4.2",
    "googlesearch-python>=1.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-docx>=1.2.0",
    "serpapi>=0.1.5",
    "xlsxwriter>=3.2.5",
]
